requirements.txt
click==6.6
Flask==0.11.1
futures==3.0.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-dateutil==2.5.3
python-swiftclient==3.0.0
requests==2.10.0
six==1.10.0
Werkzeug==0.11.10
